Use GitHub actions to automatically build and deploy website

+++ ts/backend/script-clear-json.ts
@@ -5,8 +5,10 @@ import { Log, PATH_JSON } from './general.js';
 function main() {
     try {
         // delete json directory, and create a fresh one
-        Log.action(`Deleting '${PATH_JSON}'`);
-        fs.rmSync(PATH_JSON, { recursive: true });
+        if (fs.existsSync(PATH_JSON)) {
+            Log.action(`Deleting '${PATH_JSON}'`);
+            fs.rmSync(PATH_JSON, { recursive: true });
+        }
         Log.action(`Creating '${PATH_JSON}'`);
         fs.mkdirSync(PATH_JSON);
         Log.success('Done');
